+++ app/Livewire/Organizations/Brands/Branches/ServicePoints/Qr/Show.php
@@ -160,7 +160,7 @@ class Show extends Component
         $user = $this->currentUser();
 
         if (
-            ! $user->canAccessOrganization($this->organization)
+            ! $user->canAccessBranch($this->branch, $this->organization)
             || ! $user->hasPermission(SystemPermission::GenerateQr, $this->organization)
         ) {
             abort(403);